Support dynamic import of Mediapipe
