+++ rollup.config.js
@@ -89,7 +89,6 @@ export default [
       format: 'esm',
       banner: bannerText,
       sourcemap: true,
-      inlineDynamicImports: true,
     },
     plugins: [
       typescript({
@@ -117,7 +116,6 @@ export default [
       file: 'build/xrblocks.min.js',
       format: 'esm',
       sourcemap: true,
-      inlineDynamicImports: true,
     },
     plugins: [
       typescript({
